Alpha version minus tutorials
It's done! Alpha version has been implemented. A lot of the code is
still really messy, but it works well. No stress testing yet.

+++ TheGreatPaperAdventure/Assets/Scripts/GameOverScript.cs
@@ -24,7 +24,7 @@ public class GameOverScript : MonoBehaviour
 
 	string convertScoreToDisplay(float score)
 	{
-		return (string.Format ("{0:00}:{1:00}:{2:00}", (score / 60), (int)(score % 60), (score - (float)Mathf.Floor (score))*100) );
+		return (string.Format ("{0:00}:{1:00}:{2:00}", ((int)score / 60), (int)(score % 60), (score - (float)Mathf.Floor (score))*100) );
 	}
 
 	/// <summary>
